Dedupe budget-amount formatting into shared helper
fmtAmount (LimitReachedDialog) and fmtUsage (UsageSection) were identical
copies of the same tokens/cost/messages formatter. Extract fmtBudgetAmount
into lib/format.ts and use it in both, dropping LimitReachedDialog's private
fmtTokens copy in the process.
